Adding the junit option
This uses existing code from ruff and "just works".
The test cases output are okay, but they are missing
bits of data that would be useful, and would normally
be present in the default output format.

I'm not sure what approach I will take from here, but
I think, if it's possible, it would be good to have the
entire text message that would normally appear, in the
failure message of the junit test cases. Any decent junit
report parser should display these well, in places such as
the results of a CI job.