[ty] Fix find references for type defined in stub
ty failed to return any references for a type defined in a stub where
there's also a definition in a non-stub file.

The underlying issue was that `find_references` resolved the name to
"go-to definition" whereas the visitor resolved the declaration, which
can resolve to a different file and location.

This fix uses go to declaration in both steps. I don't think it really
matters whether we use go to declaration or go to definition because
both should resolve to the same references. I picked go to declaration
because it requires less work (no stub mapping)