add unstable test case
the first of these is unstable because it formats as:

```py
(
    lambda # comment 2
    *x: x
)
```

without the second space before the comment. As a leading comment on the `*x`
parameter, it should instead format like:

```py
(
    lambda
	# comment 2
    *x: x
)
```

which is also consistent with its formatting in the presence of an additional
dangling comment on the lambda itself. this is the second test case added here,
which formats (stably) as:

```py
(
    lambda  # comment 1
    # comment 2
    *x: x
)
```