Fix comment placement in lambda parameters
Summary
--

This PR makes two changes to comment placement in lambda paramters. First, we
now insert a line break if the first parameter has a leading comment:

```py
(
    lambda
    * # comment 2
    x:
    x
)

(
    lambda # comment 2
    *x: x
)

(
    lambda
	# comment 2
    *x: x
)
```

Note the missing space in the output from main. This case is currently unstable
on main. Also note that the new formatting is more consistent with our stable
formatting in cases where the lambda has its own dangling comment:

```py
(
    lambda # comment 1
    * # comment 2
    x:
    x
)

(
    lambda  # comment 1
    # comment 2
    *x: x
)
```

and when a parameter without a comment precedes the split `*x`:

```py
(
    lambda y,
    * # comment 2
    x:
    x
)

(
    lambda y,
    # comment 2
    *x: x
)
```

Second, this PR modifies the comment placement such that `# comment 2` in these
outputs is still a leading comment on the parameter. This is also not the case
on main, where it becomes a dangling lambda comment. This doesn't cause any
instability that I'm aware of on main, but it does cause problems when trying to
adjust the placement of dangling lambda comments in #21385. Changing the
placement in this way should not affect the formatting here.

Test Plan
--

New lambda tests, plus existing tests covering the cases above with multiple
comments around the parameters (see lambda.py 122-143, and 122-205 or so more
broadly)